Split authOptions into its own file
