+++ src/app/api/fruit_locations/route.ts
@@ -1,7 +1,7 @@
 import { sql } from '@vercel/postgres';
 import { NextResponse } from 'next/server';
 import { getServerSession } from "next-auth/next";
-import { authOptions } from "../auth/[...nextauth]/route";
+import { authOptions } from "../auth/authOptions";
 
 /**
  * GET endpoint for fetching fruit tree location data. All properties are URL search parameters
